# api/openapi.yaml
openapi: 3.1.0
info:
  title: Noventrix StayOS API
  version: 1.0.0
  description: "# Noventrix StayOS - Comprehensive Hospitality Management API\n\n**Version:** 1.0.0  \n**Status:** Production\
    \ Ready  \n**Last Updated:** 2025-12-18\n\n## Overview\n\nStayOS provides a complete RESTful API for hospitality management,\
    \ covering:\n- Property & Inventory Management\n- Reservation & Booking Lifecycle\n- Guest Profiles & Loyalty\n- Operations\
    \ (Housekeeping & Maintenance)\n- Billing & Payments\n- Compliance & Audit\n- Channel Management & Distribution\n\n##\
    \ Architecture Principles\n\n- **RESTful Design:** Resource-oriented URLs, HTTP verbs, standard status codes\n- **HATEOAS:**\
    \ Hypermedia links for discoverability\n- **Idempotency:** Safe retries using idempotency keys\n- **Versioning:** URL-based\
    \ versioning (/v1/, /v2/)\n- **Pagination:** Cursor-based for large datasets\n- **Filtering:** Rich query parameters following\
    \ JSON:API conventions\n- **Rate Limiting:** Token bucket algorithm (100 req/min standard, 1000 req/min premium)\n\n##\
    \ Authentication & Security\n\n- **OAuth 2.0 + OpenID Connect** for user authentication\n- **API Keys** for server-to-server\
    \ integration\n- **JWT Bearer Tokens** with short expiry (15 minutes)\n- **Refresh Tokens** for session management\n-\
    \ **Scopes** for fine-grained authorization\n- **mTLS** for high-security integrations\n\n## Error Handling\n\nAll errors\
...
  contact:
...
  license:
...
servers:
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      description: OAuth 2.0 authorization with OpenID Connect
      flows:
        authorizationCode:
          authorizationUrl: https://auth.stayos.com/oauth/authorize
          tokenUrl: https://auth.stayos.com/oauth/token
          refreshUrl: https://auth.stayos.com/oauth/refresh
          scopes:
            read:properties: Read property information
            write:properties: Create and update properties
            read:inventory: Read inventory units
            write:inventory: Manage inventory units
            read:reservations: View reservations
            write:reservations: Create and modify reservations
            read:guests: View guest profiles
            write:guests: Manage guest profiles
            read:stays: View active stays
            write:stays: Manage check-in/check-out
            read:operations: View housekeeping and maintenance
            write:operations: Manage operations tasks
            read:payments: View payment information
            write:payments: Process payments
            read:channels: View channel manager integrations
            write:channels: Manage channel connections
            read:fb: View food & beverage orders
            write:fb: Manage F&B operations
            read:billing: View billing and folios
            write:billing: Manage billing operations
            read:procurement: View purchase orders and suppliers
            write:procurement: Manage procurement
            read:webhooks: View webhook configurations
            write:webhooks: Manage webhooks
            read:compliance: View compliance reports
            read:system: View system metrics
            read:audit: Access audit logs
            write:compliance: Generate compliance reports
            admin:all: Full administrative access
        clientCredentials:
          tokenUrl: https://auth.stayos.com/oauth/token
...
  parameters:
    PageSize:
      $ref: ./parameters/common-parameters.yaml#/PageSize
    PageCursor:
      $ref: ./parameters/common-parameters.yaml#/PageCursor
    PropertyId:
      $ref: ./parameters/common-parameters.yaml#/PropertyId
    SortBy:
      $ref: ./parameters/common-parameters.yaml#/SortBy
    SortOrder:
      $ref: ./parameters/common-parameters.yaml#/SortOrder
    Expand:
      $ref: ./parameters/common-parameters.yaml#/Expand
    IdempotencyKey:
      $ref: ./parameters/common-parameters.yaml#/IdempotencyKey
    IfMatch:
      $ref: ./parameters/common-parameters.yaml#/IfMatch
    IfNoneMatch:
      $ref: ./parameters/common-parameters.yaml#/IfNoneMatch
  headers:
    X-RateLimit-Limit:
      description: Request limit per time window
      schema:
        type: integer
    X-RateLimit-Remaining:
      description: Remaining requests in current window
      schema:
        type: integer
...
    X-Request-Id:
      description: Unique request identifier for tracing
      schema:
        type: string
        format: uuid
    ETag:
      $ref: ./headers/common-headers.yaml#/ETag
    Location:
      $ref: ./headers/common-headers.yaml#/Location
  responses:
    NoContent:
      $ref: ./responses/common-responses.yaml#/NoContent
    BadRequest:
      $ref: ./responses/common-responses.yaml#/BadRequest
    Unauthorized:
      $ref: ./responses/common-responses.yaml#/Unauthorized
    Forbidden:
      $ref: ./responses/common-responses.yaml#/Forbidden
    NotFound:
      $ref: ./responses/common-responses.yaml#/NotFound
    Conflict:
      $ref: ./responses/common-responses.yaml#/Conflict
    PreconditionFailed:
      $ref: ./responses/common-responses.yaml#/PreconditionFailed
    TooManyRequests:
      $ref: ./responses/common-responses.yaml#/TooManyRequests
    InternalServerError:
      $ref: ./responses/common-responses.yaml#/InternalServerError
  schemas:
    ProblemDetails:
      allOf:
        - $ref: ./schemas/common.yaml#/ProblemDetails
    PaginatedResponse:
      allOf:
        - $ref: ./schemas/common.yaml#/PaginatedResponse
...
    Property:
      allOf:
        - $ref: ./schemas/property.yaml#/Property
...
    UnitType:
      allOf:
        - $ref: ./schemas/property.yaml#/UnitType
...
paths:
  /properties:
    get:
      tags:
      - Properties
      summary: List properties
      description: 'Retrieve a paginated list of properties. Supports filtering, sorting, and field expansion.


        **Authorization:** Requires `read:properties` scope


        **Rate Limit:** 100 requests/minute

        '
      operationId: listProperties
      security:
      - OAuth2:
        - read:properties
      - BearerAuth: []
      parameters:
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/PageCursor'
      - name: organizationId
        in: query
        description: Filter by organization
        schema:
          type: string
          format: uuid
      - name: operationalStatus
        in: query
        description: Filter by operational status
        schema:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - CLOSED
      - name: classification
        in: query
        description: Filter by property classification
        schema:
          type: string
          enum:
          - HOTEL
          - HOSTEL
          - RESORT
          - BOUTIQUE
      - $ref: '#/components/parameters/SortBy'
      - $ref: '#/components/parameters/SortOrder'
      - $ref: '#/components/parameters/Expand'
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/PaginatedResponse'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/Property'
              example:
                data:
                - propertyId: 550e8400-e29b-41d4-a716-446655440000
                  name: Grand Plaza Hotel
                  classification: HOTEL
                  starRating: 4.5
                  operationalStatus: ACTIVE
                  timezone: America/Los_Angeles
                  location:
                    city: San Francisco
                    country: US
                  links:
                    self: https://api.stayos.com/v1/properties/550e8400-e29b-41d4-a716-446655440000
                pagination:
                  pageSize: 20
                  cursor: eyJpZCI6MTIzNDU2fQ==
                  hasMore: true
                links:
                  self: https://api.stayos.com/v1/properties?pageSize=20
                  next: https://api.stayos.com/v1/properties?pageSize=20&cursor=eyJpZCI6MTIzNDU2fQ==
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
      - Properties
      summary: Create property
      description: 'Create a new property. This is typically an administrative operation.


        **Authorization:** Requires `write:properties` or `admin:all` scope


        **Idempotency:** Supports idempotency via `Idempotency-Key` header

        '
      operationId: createProperty
      security:
      - OAuth2:
        - write:properties
        - admin:all
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Property'
            example:
              name: Grand Plaza Hotel
              classification: HOTEL
              starRating: 4.5
              location:
                street: 123 Main Street
                city: San Francisco
                state: CA
                postalCode: '94102'
                country: US
              coordinates:
                latitude: 37.7749
                longitude: -122.4194
              timezone: America/Los_Angeles
              policies:
                checkInTime: '15:00'
                checkOutTime: '11:00'
                cancellationPolicy:
                  refundability: FULLY_REFUNDABLE
                  freeCancellationHours: 24
              amenities:
              - WIFI
              - PARKING
              - POOL
              - GYM
              contactInfo:
                email: info@grandplaza.com
                phone: '+14155551234'
      responses:
        '201':
          description: Property created successfully
          headers:
            Location:
              $ref: '#/components/headers/Location'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /properties/{propertyId}:
    get:
      tags:
      - Properties
      summary: Get property by ID
      description: Retrieve detailed information about a specific property
      operationId: getProperty
      security:
      - OAuth2:
        - read:properties
      - BearerAuth: []
      parameters:
      - $ref: '#/components/parameters/PropertyId'
      - $ref: '#/components/parameters/Expand'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Successful response
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '304':
          description: Not Modified (cached version is current)
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
      - Properties
      summary: Replace property (full update)
      description: 'Replace an existing property with a complete new representation.

        All fields must be provided. For partial updates, use PATCH instead.


        **Optimistic Locking:** Use `If-Match` header with ETag to prevent conflicts

        '
      operationId: replaceProperty
      security:
      - OAuth2:
        - write:properties
      parameters:
      - $ref: '#/components/parameters/PropertyId'
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Property'
      responses:
        '200':
          description: Successfully replaced
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
      - Properties
      summary: Update property (partial update)
      description: "Partially update a property. Only provided fields will be updated.\nThis is the recommended method for\
        \ most updates.\n\n**Optimistic Locking:** Use `If-Match` header with ETag to prevent conflicts\n\n**Example:** Update\
        \ only the property name and contact email:\n```json\n{\n  \"name\": \"New Property Name\",\n  \"contactInfo\": {\n\
        \    \"email\": \"newemail@property.com\"\n  }\n}\n```\n"
      operationId: updateProperty
      security:
      - OAuth2:
        - write:properties
      parameters:
      - $ref: '#/components/parameters/PropertyId'
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Partial property update (only include fields to change)
            example:
              name: Updated Property Name
              contactInfo:
                email: newemail@example.com
      responses:
        '200':
          description: Successfully updated
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
      - Properties
      summary: Delete property
      description: 'Soft delete a property. Property status changes to CLOSED.


        **Warning:** This operation cannot be undone. All future reservations will be cancelled.

        '
      operationId: deleteProperty
      security:
      - OAuth2:
        - admin:all
      parameters:
      - $ref: '#/components/parameters/PropertyId'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot delete property with active reservations
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /properties/{propertyId}/unit-types:
    get:
      tags:
        - Inventory
      summary: List unit types
      description: Retrieve all unit types for a property
      operationId: listUnitTypes
      security:
        - OAuth2:
            - read:inventory
      parameters:
        - $ref: '#/components/parameters/PropertyId'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/UnitType'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Inventory
      summary: Create unit type
      description: Create a new unit type for a property
      operationId: createUnitType
      security:
        - OAuth2:
            - write:inventory
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitType'
      responses:
        '201':
          description: Unit type created
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
security:
- OAuth2:
  - read:properties
  - write:properties
  - read:reservations
  - write:reservations
  - read:guests
  - write:guests
  - read:payments
  - write:payments
  - admin:all
- BearerAuth: []
- ApiKeyAuth: []